+++ include/mainwindow.h
@@ -123,6 +123,8 @@ public:
     Q_INVOKABLE void setMetatileTerrainType(int metatileId, int terrainType);
     Q_INVOKABLE int getMetatileBehavior(int metatileId);
     Q_INVOKABLE void setMetatileBehavior(int metatileId, int behavior);
+    Q_INVOKABLE QString getMetatileBehaviorName(int metatileId);
+    Q_INVOKABLE void setMetatileBehaviorName(int metatileId, QString behavior);
     Q_INVOKABLE int getMetatileAttributes(int metatileId);
     Q_INVOKABLE void setMetatileAttributes(int metatileId, int attributes);
     Q_INVOKABLE QJSValue getMetatileTile(int metatileId, int tileIndex);